pyproject.toml
[tool.poetry]
name = "ella"
version = "0.1.0"
description = ""
authors = ["Jade Wang <jadewang@umich.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pylint = "^3.3.2"
mypy = "^1.13.0"
isort = "^5.13.2"
black = "^24.10.0"
ipdb = "^0.13.13"
pytest = "^8.3.4"
torch = "2.2.2"
numpy = "^1.0"
scipy = "^1.0"
matplotlib = "^3.9.3"
lightning = "^2.4.0"
tensorboard = "^2.18.0"
hydra-core = "^1.3.2"
pydantic = "^2.10.4"
pylint-pydantic = "^0.3.4"
ipykernel = "^6.29.5"
shapely = "^2.0.6"
pandas = "^2.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ella-train = "ella.cli.train:main"
ella-estimate = "ella.cli.estimate:main"
ella-visualize = "ella.cli.visualize:main"

[tool.black]
line-length = 119

[tool.isort]
profile = "black"
